[HGXB300]Memory corruption issue
* issue: MCTPD earlier used to assume Bridge's own eid < Bridge pool
  start. Something which is not explicity said but is interpreted from

DSP0236 : 9.1.9 Consolidating routing table entries

MCTP requires that when an EID pool is allocated to a device, the range of EIDs is contiguous and follows the EID for the bridge itself.

But with current HGXB300 configuration, some of MCU's eid > their pool
start. This caused array out of bound update on local dynamic arrays
causing heap memory corruption.

* fix: avoid bridge's own eid from routing table entry

Tested on HGXB300

Fixes JIRA https://

Signed-off-by: Faizan Ali <faizana@nvidia.com>